+++ src/components/ui/typography/MarkdownRenderer/inlineStyles.tsx
@@ -103,7 +103,7 @@ const DeletedText = ({ className, ...props }: DeletedTextProps) => (
 
 const HorizontalRule = ({ className, ...props }: HorizontalRuleProps) => (
   <hr
-    className={cn(className, 'mx-auto w-48 border-b border-border')}
+    className={cn(className, 'border-border mx-auto w-48 border-b')}
     {...filterProps(props)}
   />
 )
@@ -156,7 +156,7 @@ const Img = ({ src, alt }: ImgProps) => {
   return (
     <div className="w-full max-w-xl">
       {error ? (
-        <div className="flex h-40 flex-col items-center justify-center gap-2 rounded-md bg-secondary/50 text-muted">
+        <div className="bg-secondary/50 text-muted flex h-40 flex-col items-center justify-center gap-2 rounded-md">
           <Paragraph className="text-primary">Image unavailable</Paragraph>
           <Link
             href={src}
@@ -183,7 +183,7 @@ const Img = ({ src, alt }: ImgProps) => {
 
 const InlineCode = ({ children }: { children?: React.ReactNode }) => {
   return (
-    <code className="relative whitespace-pre-wrap rounded-md bg-background-secondary/60 px-1.5 py-0.5 text-xs font-medium text-foreground border border-border/10">
+    <code className="bg-background-secondary/60 text-foreground border-border/10 relative whitespace-pre-wrap rounded-md border px-1.5 py-0.5 text-xs font-medium">
       {children}
     </code>
   )